LPS-147890 Move test cases
